Split Parsers.swift
